Lockers

+++ folders.php
@@ -3,25 +3,30 @@
 require_once "functions.php";
 if (!$u)
     diez();
-if (!array_key_exists("eid",$req))
-    diez();
-$eid = $req['eid'];
-$a = UserAccessEP($eid,$u->uid);
+$lid = 0;
+$eid = 0;
+if (array_key_exists("eid",$req))
+    $eid = $req['eid'];
+if (array_key_exists("lid",$req))
+    $lid = $req['lid'];
+if ($eid)
+    $a = UserAccessEP($eid,$u->uid);
+if ($lid)
+    $a = UserAccessLocker($lid,$u->uid);
 if ($a != 2)
     diez();
 
 require_once "output.php";
 
 
-
 if (array_key_exists("delete",$req))
 {
-    if (array_key_exists("force",$req))
+    if (array_key_exists("force",$req) && UserAccessFolder($req['delete'],$u->uid) == 2)
     {
         QQ("BEGIN TRANSACTION");
         DeleteWholeFolder($req['delete']);
         QQ("COMMIT");
-        redirect(sprintf("folders.php?eid=%s",$eid));
+        redirect(sprintf("folders.php?eid=%s&lid=%s",$eid,$lid));
     }
     echo '<div class="content" style="margin:20px;">';
     echo 'Διαγραφή Φακέλου';
@@ -30,8 +35,8 @@ if (array_key_exists("delete",$req))
 
     if ($docccount)
         printf("<br><b>Η διαγραφή του φακέλου θα διαγράψει οριστικά και %s έγγραφα που βρίσκονται σε αυτόν!</b>",$docccount);
-    printf('<br><br><button href="folders.php?eid=%s&delete=%s&force=1" class="button autobutton is-danger">Επιβεβαίωση Διαγραφής</button><br>',$req['eid'],$req['delete']);
-    printf('<br><hr><button href="folders.php?eid=%s" class="button autobutton is-success">Πίσω</button>',$req['eid']);
+    printf('<br><br><button href="folders.php?eid=%s&lid=%s&delete=%s&force=1" class="button autobutton is-danger">Επιβεβαίωση Διαγραφής</button><br>',$eid,$lid,$req['delete']);
+    printf('<br><hr><button href="folders.php?eid=%s&lid=%s" class="button autobutton is-success">Πίσω</button>',$eid,$lid);
     die;
 }
 
@@ -45,16 +50,16 @@ if (array_key_exists("c",$_POST))
     if (!$r)
     {
         // Create
-        QQ("INSERT INTO FOLDERS (EID,NAME,PARENT,SPECIALID,CLASSIFIED) VALUES (?,?,?,?,?)",array($_POST['eid'],$_POST['name'],$_POST['parent'],0,$_POST['classification']));
+        QQ("INSERT INTO FOLDERS (EID,LID,NAME,PARENT,SPECIALID,CLASSIFIED) VALUES (?,?,?,?,?,?)",array($_POST['eid'],$_POST['lid'],$_POST['name'],$_POST['parent'],0,$_POST['classification']));
     }
     else
     {
         // Edit
         if ($_POST['parent'] == $_POST['c'])
             $_POST['parent'] = 0;
-        QQ("UPDATE FOLDERS SET EID = ?, NAME = ?, PARENT = ?, CLASSIFIED = ? WHERE ID = ? AND SPECIALID = 0",array($_POST['eid'],$_POST['name'],$_POST['parent'],$_POST['classification'],$_POST['c']));
+        QQ("UPDATE FOLDERS SET EID = ?,LID = ?, NAME = ?, PARENT = ?, CLASSIFIED = ? WHERE ID = ? AND SPECIALID = 0",array($_POST['eid'],$_POST['lid'],$_POST['name'],$_POST['parent'],$_POST['classification'],$_POST['c']));
     }
-    redirect(sprintf("folders.php?eid=%s",$eid));
+    redirect(sprintf("folders.php?eid=%s&lid=%s",$eid,$lid));
     die;
 }
 
@@ -62,21 +67,23 @@ function CreateOrEditFolder($fid)
 {
     global $u;
     global $eid;
+    global $lid;
     $r = QQ("SELECT * FROM FOLDERS WHERE ID = ?",array($fid))->fetchArray();
     if (!$r)
-        $r = array("ID" => 0,"NAME" => "", "PARENT" => "","CLASSIFIED" => "");
+        $r = array("ID" => 0,"NAME" => "", "PARENT" => "","CLASSIFIED" => "","LID" => 0);
 
     ?>
     <form method="POST" action="folders.php">
     <input type="hidden" name="c" value="<?= $r['ID'] ?>" />
     <input type="hidden" name="eid" value="<?= $eid ?>" />
+    <input type="hidden" name="lid" value="<?= $lid ?>" />
 
     Όνομα: <br>
         <input type="text" class="input" name="name" value="<?= $r['NAME']?>" required/>
     <br><br>
     Μέσα σε: <br>
     
-    <?= PickFolder($eid,"parent",array($r['PARENT'])) ?>
+    <?= PickFolder($eid,"parent",array($r['PARENT']),0,0,0,0,$lid) ?>
 
     <br><br>
     Διαβάθμιση: <br>
@@ -93,11 +100,22 @@ function CreateOrEditFolder($fid)
 
 function PrintFolders($rootint = 0,$nest = 0)
 {
-    global $eid;
-    if ($rootint == 0)
-        $q = QQ("SELECT * FROM FOLDERS WHERE EID = ? AND (PARENT = 0 OR PARENT IS NULL)",array($eid));
+    global $eid,$lid;
+    if ($lid)
+    {
+        if ($rootint == 0)
+            $q = QQ("SELECT * FROM FOLDERS WHERE LID = ? AND (PARENT = 0 OR PARENT IS NULL)",array($lid));
+        else
+            $q = QQ("SELECT * FROM FOLDERS WHERE LID = ? AND PARENT = ?",array($lid,$rootint));
+    }
     else
-        $q = QQ("SELECT * FROM FOLDERS WHERE EID = ? AND PARENT = ?",array($eid,$rootint));
+    {
+        if ($rootint == 0)
+            $q = QQ("SELECT * FROM FOLDERS WHERE EID = ? AND (PARENT = 0 OR PARENT IS NULL)",array($eid));
+        else
+            $q = QQ("SELECT * FROM FOLDERS WHERE EID = ? AND PARENT = ?",array($eid,$rootint));
+
+    }
     ?>
     <?php
     if ($nest == 0)
@@ -121,8 +139,8 @@ function PrintFolders($rootint = 0,$nest = 0)
             printf('%s',$r['NAME']);
             if ($r['SPECIALID'] == 0)
             {
-                printf(' <a href="folders.php?eid=%s&fid=%s"><font color="red">Επεξεργασία</font></a>',$eid,$r['ID']);
-                printf(' <a href="folders.php?eid=%s&delete=%s"><font color="red">Διαγραφή</font></a>',$eid,$r['ID']);
+                printf(' <a href="folders.php?eid=%s&fid=%s&lid=%s"><font color="red">Επεξεργασία</font></a>',$eid,$r['ID'],$lid);
+                printf(' <a href="folders.php?eid=%s&delete=%s&lid=%s"><font color="red">Διαγραφή</font></a>',$eid,$r['ID'],$lid);
             }
             PrintFolders($r['ID'],$nest + 1);
             printf('</td>');
@@ -134,8 +152,8 @@ function PrintFolders($rootint = 0,$nest = 0)
             for($i = 0 ; $i < $nest ; $i++)
                 printf('&nbsp;');
             printf('(%s) %s',$r['ID'],$r['NAME']);
-            printf(' <a href="folders.php?eid=%s&fid=%s"><font color="red">Επεξεργασία</font></a>',$eid,$r['ID']);
-            printf(' <a href="folders.php?eid=%s&delete=%s"><font color="red">Διαγραφή</font></a>',$eid,$r['ID']);
+            printf(' <a href="folders.php?eid=%s&fid=%s&lid=%s"><font color="red">Επεξεργασία</font></a>',$eid,$r['ID'],$lid);
+            printf(' <a href="folders.php?eid=%s&delete=%s&lid=%s"><font color="red">Διαγραφή</font></a>',$eid,$r['ID'],$lid);
             PrintFolders($r['ID'],$nest + 1);
         }
     }
@@ -156,7 +174,7 @@ if (array_key_exists("fid",$req))
 }
 else
 {
-    PrintHeader('endpoints.php',sprintf('&nbsp; <button class="button is-primary is-small autobutton" href="folders.php?fid=0&eid=%s">Νέος Φάκελος</button>',$req['eid']));
+    PrintHeader('endpoints.php',sprintf('&nbsp; <button class="button is-primary is-small autobutton" href="folders.php?fid=0&eid=%s&lid=%s">Νέος Φάκελος</button>',$eid,$lid));
     PrintFolders();
 }
 ?>
